+++ prisma/migrations/20260408093313_update_wind_data/migration.sql
@@ -1,2 +1,0 @@
--- CreateIndex
-CREATE INDEX "WindData_sensorId_idx" ON "WindData"("sensorId");