+++ src/configs/forecast.config.ts
@@ -1,0 +1,5 @@
+import { getEnv } from '@/shared/utils/env.util';
+
+export const forecastConfig = {
+    url: getEnv('FORECASE_URL'),
+} as const;