+++ src/modules/forecast/forecast.api-controller.ts
@@ -1,0 +1,6 @@
+import { Request, Response } from 'express';
+
+export const createForcast = (req: Request, res: Response) => {
+    // TODO: implement queue (bull)
+    res.json({ status: 'success', data: {} });
+};